xfsprogs: Add fsck.xfs cli
This cli has been rewritten in Rust as bottlerocket does not have
shell.
This CLI is intented to get called uninteractively:
1. By Kubernetes
2. By systemd-fsck-root.service - This is disabled by default in
   bottlerocket.

In case of Kubernetes, this binary will be used to repair the filesystem
when it exists.
We are keeping it open for interactive sesssion too, though it is
recommended to invoke xfs_repair for Normal administrative filesystem
repairs directly.