release: disable synthesized hostname
Otherwise, the reverse lookup in `netdog` will return "linux" if no
other name can be resolved by DNS. This causes problems when trying
to join several nodes to a Kubernetes cluster, since they all attempt
to register as "linux".

The expected (historical) behavior is that the system hostname winds
up as the primary interface's IP address in this case. This is not
ideal and causes other headaches: for example, the short name of the
host is the first part of the IP in dotted decimal notation. But in
the absence of a setting or reverse DNS, it's at least a potentially
predictable and unique identifier.

Signed-off-by: Ben Cressey <bcressey@amazon.com>