models: add new setting to auto-load kernel modules
Loading specific kernel modules can be necessary to use specific
features on your node. One example would be loading the correct module
for your choice of scheduling algorithm for ipvs.

Loading kernel modules is currently only possible during boot and requires
the use and maintenance of an init-container image to load the desired
kernel modules on boot.

We can simplify that by adding the additional setting `autoload` for
kernel modules and utilizing the already available
systemd-modules-load service.

Module auto-loading is in conflict with blocking modules from loading
through the sibling setting `allowed`. Hence, do not auto-load a module
if the module is not allowed at the same time. Blocking takes precedence,
as it is the more prohibitive operation.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>