actions: Enable list-variants to filter on changes
This teaches the list-variants GitHub action to be able to return only a
list of variants that are impacted by the changes in the PR.

It switches the list-variants action to be a JavaScript action. This
allows us to have more control over what we can do and how we process
the pull request information.

By default it will not filter any of the variants and will just return
whatever is found under the variants directory. This is useful for the
current nightly workflow and if we add anything else that needs to do
anything for all variants included.

There is now an input called filter-variants that enables the new
behavior. If that is set to true, the action will get the list of
variants from the repo, but also try to determine all paths in the repo
that make up each variant. With that information, it is able to look at
the files being modified as part of the PR and determine which variants
are actually impacted by the change. Any variants that do not appear to
be modified will be filtered out of the returned list, so that the build
workflow will only include affected variants as part of the matrix of
tests to run.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>