actions: Add action file hash to cache identifier
When caching was added, it was all performed inside the cache.yml
workflow file. These have since been refactored and the caching is now
done in a composite action that is used by multiple worflows. This means
the hashing that was originally done on the cache.yml workflow file is
no longer enough to ensure the uniqueness when changes are made.

This adds the hash of all *.yml (worflows and actions) to make sure
changes are picked up for cache identification.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>