POC: Filter workflow variants on changed files
This uses a combination of the get-changed-files action and bash
scripting to determine what variants need testing based on the changes.
This would require updates to buildsys or another tool that could tell
us what paths in the repo are relevant to each variant to be able to
make the decision if it needs to be tested.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>