apiclient: update tungstenite to v0.20.1
Newer versions of tungstenite implemented the v13 version of the Web
Sockets protocol (described in https://www.rfc-editor.org/rfc/rfc6455).
In this version, new headers are required in the upgrade request. With
this update, now the upgrade request sent by apiclient include the
missing headers.

In the same version, the 'frame' message was added. Neither apiserver
nor apiclient send this type of message, so they are ignored by
apiclient.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>