release: Remove explicit Type for Local Filesystem
The filesystem will mount without specifying a specific type when
defining a systemd Mount. This allows any filesystem that the OS can
mount to be used as a filesystem for /local. Before this change, the
boot would fail if the /local filesystem didn't match the one specified
in /usr/share/bottlerocket/image-features.env

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>