packages: Update open-vm-tools to 12.3.0
This updates the `open-vm-tools` package to use the 12.3.0 release.

There are a lot of merged changes since the previous release [0], but
most are called out as not being applicable for this package. The only
significant change appears to be the commit to add addition x509 checks
to the SAM token verification.

[0] https://github.com/vmware/open-vm-tools/compare/stable-12.2.5...stable-12.3.0
[1] https://github.com/vmware/open-vm-tools/commit/74b6d0d9000eda1a2c8f31c40c725fb0b8520b16

Signed-off-by: Sean McGinnis <stmcg@amazon.com>
(cherry picked from commit f944f80329b5f85eeba9048d8fb4b80e2ee9ef14)