systemd: Backport patches that allow netlink timeout to be configured
Backport 2 patches that allow the configurability of the default timeout
for waiting for replies for netlink messages from the kernel.  The
default timeout is 25 seconds and it is possible to hit this timeout if
the system is under load.  If the timeout is hit, network links can
become unusable without intervention.

Reference issue: github.com/systemd/systemd/issues/25441