selinux-policy: allow fanotify permission events
fanotify supports permission events that allow the watching process
to determine whether a file can be accessed, opened, or executed. If
a process sets these event flags when marking a filesystem object,
then SELinux allows us to deny that access. However, once the mark is
set, the process is free to prevent any other process from accessing
the watched files. These denials are not logged by the kernel and may
not be logged at all, leading to failures that would be difficult to
debug.

For that reason, the related actions are only allowed for "trusted"
subjects. Since the host doesn't currently make use of fanotify, in
practice this just means containers running with the "super_t" label.
The goal is to enable vendors who are aware of the pitfalls to use
this functionality in a way that does not interfere with the expected
operation of the host.

Signed-off-by: Ben Cressey <bcressey@amazon.com>