bloodhound: Fix journal file permission check
The Bottlerocket CIS check 4.1.2 would correctly report if there was a
failing condition, but it missed changing the result from SKIP to PASS
if there was no failure.

This updates the result for 4.1.2 so that it will correctly report
success as long as there are no non-compliant journal file permissions.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>