bloodhound: Improve chronyd active check
The existing check for chronyd is fine if the service is running and it
properly recognizes the state. If the service is stopped or failed
though, the check directs the user to perform a manual check due to the
way the status is being checked. The "systemctl is-active" call returns
a failure exit code, causing the check to fail to determine its state.

This changes the way the check is performed by calling "systemctl show
--property ActiveState chronyd". This command will just report the
state, and its exit status will be 0 as long as the state can be
retrieved. It is not affected by what that state actually is. The string
check of the output is also improved to explicitly look for the full
ActiveState output.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>