Change nightly GitHub Action to run weekly
The nightly job was added to help catch cases where yanked Go modules
would be found right away by using the GOPROXY=direct setting to bypass
any caching.

This has been in place for a few months, and so far the only failures
seen have been due to Docker repo throttling (too many requests). Since
this job runs every variant in the repo, it is a relatively expensive
job to be running every night.

This switches the job from nightly to weekly. This will still give an
early warning in case of any yanked modules, but will limit the amount
of overall runtime that is usually unnecessary.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>