cis: Improve SCTP kernel module check
The level 2 check 3.3.1 verifies the sctp kernel module is disabled. The
current check for whether it was already loaded was not correct. There
is a second check to make sure loading is also disabled. If the setting
for loading is to not allow it, but the module is already laoded, the
check does not return the expected output. So it would stillr eport
failure, but it is less than ideal for reporting the actual issue.

This changes the check for whether the module is loaded to correctly
identify if the module is loaded or not before checking whether the
ability has been disabled.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>
(cherry picked from commit d411f603dda549ddafaaf2ed1dbe026f97f95e3b)