packages: Add 'early-boot-config-local' package
Adds a new package for the `local-user-data` provider.  The package
includes 2 subpackages, a '*-client' package which includes just the
binary, and a '*-provider' package which symlinks the binary to the .d
directory in which `early-boot-config` will look for providers.  The
default prefix in the '*-provider' package is '20-' to place it second
in order.