packages: Add 'early-boot-config-vmware-cd-rom' package
Adds a new package for the `vmware-cd-rom` provider.  The package
includes 2 subpackages, a '*-client' package which includes just the
binary, and a '*-provider' package which symlinks the binary to the .d
directory in which `early-boot-config` will look for providers.  The
default prefix in the '*-provider' package is '30-' to place it third in
order.