early-boot-config: Add EC2 IMDS user data provider
Adds a standalone user data provider binary specifically for EC2 IMDS.
The binary is meant to be run by the main `early-boot-config`
executable.  It logs to stderr, as it returns a `SettingsJson` object on
stdout in the case of success.  It returns an error and non-zero exit
code in the case of failure.