early-boot-config: Add local overrides user data provider
Adds a standalone data provider binary specifically for the local
overrides file `/local/user-data-overrides.toml`.  The binary is meant
to be run by the main `early-boot-config` executable.  It logs to
stderr, as it returns a `SettingsJson` object on stdout in the case of
success.  It returns an error and non-zero exit code in the case of
failure.