release: mount /opt/cni rather than /opt/cni/bin
This is done for consistency with the new /opt/csi mount, where the
helpers may need to store non-executable files as well as binaries.

Note that /opt/cni has always been cleaned up on every boot, so this
will not remove any files that weren't previously removed.

The main change is that files outside of /opt/cni/bin will now be
labeled with "cni_exec_t" instead of "local_t". These types are
largely equivalent in the current policy, in terms of file-related
permissions, so the change should be safe.

Signed-off-by: Ben Cressey <bcressey@amazon.com>