selinux-policy: add rules for CSI helpers
CSI drivers that mount filesystems with FUSE need to ensure that the
mounting process survives a container restart; otherwise, they cannot
be updated without triggering a filesystem failure in any pod which
uses the mount. One workaround for this is to have the host run the
mounting process on behalf of the container, so that the lifecycles
of the driver and the mount are no longer the same.

In some ways this is similar to CNI, where containers can provide
plugins that the host runs while setting up new network namespaces.
It's also different in that CSI mount helpers must run before the
container is created, rather than during creation. CSI mount helpers
may also need access to credentials or other secrets to perform the
mount, so the processes must be treated as privileged rather than
unprivileged containers.

Signed-off-by: Ben Cressey <bcressey@amazon.com>