selinux-policy: permit systemd to run CSI helpers
Ideally, CSI drivers that want the host to run a helper process on
their behalf would arrange to run that process inside a container, to
avoid any dependencies on host software beyond the systemd interface.
However, this isn't strictly required, and treating the process as a
container fulfills the policy objective.

Allow systemd to execute such processes directly, without requiring
them to be wrapped by a `runc` invocation.

Note however that it requires a high level of privilege to interact
with systemd via its DBUS API to create a unit and arrange for it to
run. There are no plans to relax this restriction.

Signed-off-by: Ben Cressey <bcressey@amazon.com>