selinux-policy: clarify local storage rules
The distinction between "protected" (i.e. "write-restricted") and
"restricted" (i.e. "read-restricted") was unclear and the attribute
names did not imply that one was related to the other.

Clarify this by renaming the attributes and defining the subset
relationship between them.

Another distinction in the policy is between local files that can be
mutated by both confined and unconfined system processes, and local
files that can only be mutated by unconfined system processes. The
first kind of objects are matched by rules to specific subjects; the
second kind are instead defined by the absence of rules that would
allow confined subjects to mutate them.

Add the "sensitive" attribute to collect these types and to clarify
the policy objective: these are files that can't be mutated by
containers and also can't be mutated by confined system processes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>