host-containers: use model from settings extension
Because host-containers still reads settings from the API, it reads them
to the variant's settings model, which means that now it needs to pull
in the host-containers settings extension model.

This change is only needed until host-containers is migrated to no
longer get settings directly from the API, and instead gets them from a
config file.

Signed-off-by: Sam Berning <bernings@amazon.com>