packages: update libnvidia-container to 1.14.5
Starting from version 1.14.0, upstream added the versions.mk file
that is used to create the nvc.h file from a template as part of build
process. Refer to NVIDIA/libnvidia-container@7161f94 commit for more info.

By default, the Major, Minor and Patch versions, which must be provided,
are extracted from a git tag, if available. But as we build directly from the
source, git is unavailable. Hence, set the LIB_VERSION variable in the spec file
to build the version used in nvc.h.