systemd: use bootconfig for unified cgroup hierarchy
Instead of changing the way that systemd is compiled based on the
image feature flag, the expected init parameter can be added to a
bootconfig drop-in.

The systemd package now requires one of the cgroup hierarchy packages
when installing as part of a variant image build, so that the build
will fail if neither can be installed rather than falling back to the
default behavior.

If necessary, the cgroup hierarchy setting can still be overridden at
runtime or by replacing the default bootconfig file.

Signed-off-by: Ben Cressey <bcressey@amazon.com>