schnauzer: add if_not_null template helper
There's currently not a safe way to render potentially-null falsy values
in a schnauzer template. Gating with `{{#if falsy}}` will skip
rendering, and `{{default false falsy}}` will always render something,
even if `falsy` is unset.

`{{#if_not_null falsy}}` checks only that the value is not null, it
doesn't care if it's falsy. This allows us to conditionally render
template portions that may contain falsy values.

Signed-off-by: Sam Berning <bernings@amazon.com>