add k8s-1.30 package and variants boilerplate
Adds boiler plate for k8s-1.30 variants based on previous addition of
k8s variants. This commit does not add the actual k8s-1.30 sources, but
rather sets the stage for incremental changes to k8s-1.30 variants.

This adds a k8s-1.30 package. Existing 1.29 models are moved to be 1.30,
with symlinks created for 1.29 models to point to the new locations.
For now, the k8s-1.30 package points to the k8s-1.29 source.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>