release: adjust prepare-boot service dependencies
The prepare-boot service calls `prairiedog prepare-boot`, which needs
to find the active partition set in order to mount the correct boot
filesystem.

Previously, this depended on the mark-boot-successful service, which
adjuts the GPT flags for the active partition set. However, it does
not actually change the active partition set, so it's not actually a
required dependency for prepare-boot.

This matters for FIPS integrity checks because the boot filesystem
must be mounted to verify the kernel's HMAC. If the boot is marked as
successful and then the integrity check fails, the system could not
then be rebooted back into the previous partition set.

Signed-off-by: Ben Cressey <bcressey@amazon.com>