release: add subpackage for fips
Provide overlay mounts so that binaries installed to `/usr/fips/bin`
and `/usr/fips/libexec` will be available in the standard paths.

Install a bootconfig snippet so that `fips=1` will be added to the
kernel command line, and to "systemd.unit=fipsheck.target" to systemd
to override any default target.

Ensure that `/etc/system-fips` is created by default. This file can
be mounted into containers for certain distros (RHEL, Amazon Linux)
to trigger the use of FIPS certified crypto.

Signed-off-by: Ben Cressey <bcressey@amazon.com>