release: add dependency on early-boot-config
Some form of early-boot-config is required for the OS to function, so
add that dependency to the release package, which collects the other
mandatory dependencies.

Switch to boolean dependencies in the early-boot-config package, so
that the expected providers are installed for the variant's platform.

Remove the explicit dependency from all variant definitions, as an
exercise in boilerplate reduction.

Signed-off-by: Ben Cressey <bcressey@amazon.com>