os: pessimize pluto and cfsignal builds
On a high-end build machine, `pluto` takes 12 minutes to compile when
using one codegen unit, and around 3 minutes to compile with the
default of 16 codegen units. This is because the aws-sdk-ec2 crate
contains a massive amount of code.

Sacrifice a little code quality and spec file clarity to bring build
times for this package back under control by setting up a new target
directory and another background job for `pluto`. `cfsignal` is given
the same treatment, even though aws-sdk-cloudformation is not as much
of a problem, because they share crate dependencies that aren't used
by other crates.

Refactor the handling of stdout and stderr file descriptors to ensure
that the `cargo build` invocation is shown when printing the results,
instead of when the job is first started. Otherwise, it overlaps with
the non-static, non-AWS-SDK output in a confusing way.

Signed-off-by: Ben Cressey <bcressey@amazon.com>