systemd: use unified cgroup hierarchy by default
This largely reverts bac6f6ee3cc3689ccec4bf8b0ebb1f04258f7082
("systemd: use bootconfig for unified cgroup hierarchy") because the
bootconfig snippet approach does not work for in-place upgrades, as
the bootconfig data is only consumed on first boot and then persisted
as settings.

Instead, build systemd with the "unified" hierarchy as the default.
Variants that need the "hybrid" hierarchy will have to handle that on
their own, by specifying a kernel command-line parameter.

Signed-off-by: Ben Cressey <bcressey@amazon.com>