logdog: switch to use apiclient cli to get settings
This changes logdog to fetch settings using the apiclient. This creates two breaking changes.
First logdog will now include sections such as 'os' in the log. This provides more information
in the log. Secondly instead of removing the existence of secret settings their values are replaced
with REDACTED. This allows the user to see the setting was there while maintaining the scrubbing of
secrets.