kmod-*-nvidia: specify AWS EULA as license
Previously, the NVIDIA packages were special-cased with a "bring your
own license" step, to support the case where a developer building the
project might choose the terms of a different license under which to
distribute the software or to make it available to end users.

With the advent of out-of-tree builds, packages are moving into kits
that contain pre-compiled binary RPMs, which means they will always
have a header indicating which license applies.

This commit makes it explicit what license those RPMs will have at
the time they are published.

Since the terms of the AWS EULA indicate that the software should be
used to develop AMIs for use on AWS, add an install-time requirement
for `aws-*` variants.

Signed-off-by: Ben Cressey <bcressey@amazon.com>
(cherry picked from commit c1e4d11f19f5b1a3ad3448ce1d5f4535a10fb6fb)