os: prepare for forthcoming build system changes
For now, the os package still relies on variant sensitivity to ensure
that any programs that depend on the model are built with the correct
data structure. The next version of Twoliter will support this, but
only when the compatibility feature flag is set in the environment.

Older versions of Twoliter made the intended root.json available as
a file to be packaged, while newer versions install it directly in
the final image. Include the file if it exists, otherwise skip it.

Signed-off-by: Ben Cressey <bcressey@amazon.com>