pluto: add settings to launch transaction
When settings-applier.service starts, it runs settings-committer,
because it expects that sundog will have added generated settings to
the launch transaction, and then runs thar-be-settings to apply the
changes.

However, if pluto runs, it will have already committed any settings
from sundog, and then directly applied the settings via apiclient.
This introduces a different mode for system boot depending on whether
pluto is installed, and also worsens boot time as thar-be-settings is
relatively slow.

Fix this by having pluto add its own generated settings to the launch
transaction, so that settings-committer always has pending changes to
commit, and so that thar-be-settings only runs once.

Signed-off-by: Ben Cressey <bcressey@amazon.com>