settings-plugins: do not prefer shared linking
The SDK is built for two host architectures - x86_64 and aarch64.
Each build includes the same two target architectures. However, the
libstd.so binary for a given target architecture will not be the same
across the two host SDK images, because they are built on different
systems and at different times.

To reduce file sizes, we prefer dynamically linking to libstd for all
Rust binaries, and arrange to install libstd through the libstd-rust
package. Prior to the kit migration, we always built the final image
with the same host SDK image that was used to build all packages, so
the question of which libstd should be shipped never came up.

In the kits world, we need the libstd-rust that corresponds to the
many Rust binaries that come from the core kit. However, if we prefer
dynamic linking, then we also need the libstd-rust that matches the
Rust binaries built locally, which may not be the same if the kit was
built on a different host architecture.

Sidestep this for now by not preferring shared linking for settings
plugins, so that they are statically linked to the libstd present in
the build environment, and do not require it to be installed in the
final image.

Signed-off-by: Ben Cressey <bcressey@amazon.com>