sources: move settings-sdk to workspace dependencies
Within the sources/ workspace, if a crate depends on one of the crates
in bottlerocket-settings-sdk, we currently include a section with the
full git URL, tag, and version in the crate's Cargo.toml. Because we
can't have two versions of the same crate in a workspace, updating one
of these Cargo.toml's (in order to test changes in a personal fork, or
to bump the version, for example) requires updating all of the other
crates that use that dependency.

Making the crates in bottlerocket-settings-sdk workspace dependencies
allows us to update to a different version of bottlerocket-settings-sdk
by only changing one file.

Signed-off-by: Sam Berning <bernings@amazon.com>