settings-plugins: specify plugin soname
Rust 1.81.0 has started setting the soname for cdylib artifacts it
builds, which now matches the library's name provided by the crate.

These library names could all be the same, but in practice should be
distinct within the workspace to prevent cdylib artifacts from being
overwritten when building two or more crates that share a library
name.

If a soname is not specified, then glibc's dynamic linker will treat
the library's filename as its soname. If a soname is specified, then
that will always be used regardless of the filename.

Both the settings plugin loader code and the ld.so.conf.d drop-ins
expect that we can install each crate's cdylib as "libsettings.so"
and have the dynamic linker find the right library at runtime.

To keep this working as expected, add a new linker argument so the
embedded soname ends up as "libsettings.so", to match the filename.

Signed-off-by: Ben Cressey <bcressey@amazon.com>